requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
python-telegram-bot>=13.1,<14.0
python-dotenv==0.15.0
geopy==2.1.0
pytz>=2021.1